feat: add virtualized grid to albums list page
